+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add an upload button in Dashboard Journal view #221
 - Added tutorial for User Edit and Profile view
 - Added title to charts in statistics view
+- Add multiple actions to Journal and Users view #245
 
 ### Changed
 - Change edit classroom screen to handle large number of students